hygieia/server/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
pandas==2.1.0
numpy==1.25.2
python-multipart==0.0.6
mysql-connector-python==8.1.0
passlib[argon2]
pydantic[email]
email-validator==2.2.0
scikit-learn==1.3.0
